pyproject.toml
[project]
name = "aiohttp-docs"
version = "0.0.1"
description = "API documentation for aiohttp server based on Pydantic models"
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
license-files = ["LICENSE"]
authors = [
    { name = "Mikhail Shagov", email = "mishaga@me.com" },
]
keywords = ["aiohttp", "pydantic", "swagger", "openapi", "documentation"]
classifiers = [
    "Development Status :: 1 - Planning",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "aiohttp>=3.12.15,<4.0.0",
    "pydantic>=2.11.7,<3.0.0",
]


[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-aiohttp>=1.1.0",
    "pytest-cov>=6.1.1",
    "pyyaml>=6.0.2",
    "ruff>=0.11.10",
]


[build-system]
requires = ["uv_build>=0.8.11"]
build-backend = "uv_build"


[tool.uv.build-backend]
module-name = "aiohttp_docs"
module-root = ""


[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py313"


[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"


[tool.ruff.lint]
select = ["ALL"]
ignore = ["Q000", "Q003", "COM812"]


[tool.ruff.lint.pydocstyle]
convention = "google"
